Remove redundant logic from wrap_line().
The code to wrap lines had a special check that broke up a new line
consisting of a single word when it detected the line was too
long. However, this case is already handled by the other logic in the
algorithm so there is no need for the special check.