Fix naming logic.
Previously, get_unique_name() failed to give distinct names to some
distinct input variables because it was passed the wrong
parameters. Furthermore, the same flawed naming logic was duplicated in
multiple places in the code generator.

The task of finding unique names for the Python and Fortran code
generators is moved to a new KeyToUniqueNameMap class. It has a simple
interface to query for names that are guaranteed to be distinct for
distinct input variables.

Bug: #34