Slightly relax the definition of an "If-Then-Else" control structure.
The change is that the "merge" block doesn't have to exist for a
subgraph to be identified as "If-Then-Else."

The intention of this change is to allow control flow graphs such as

    A
   / \
   B C

(where B and C exit the procedure) to be matched as If-Then-Else
nodes. This is needed to handle adaptive code.