Implement SimpleCodeBuilder.
This is an optional wrapper around CodeBuilder that is designed to
simplify the handling of dependencies and address the issue of
verbosity.

Handling of dependencies is simplified by using a block such that any
instructions added inside the block follow a linear order. In addition,
there is support for if-then style conditional dependencies.

Language verbosity is addressed by using simplified functions to add
instructions. Currently supports .assign() / .yield_state().

Bug: #30, #31