Update the multirate method code to return separated state and to avoid subscripts.
This removes internal use of subscripts and numpy object arrays in the
multirate code.

Also, update the multirate test to be more fine-grained. Previously, all
problems were run by the testing function at once. This changes the code
so that each test invocation uses a different problem. This allows for
more test granularity.

Bug: #21, #26