Progress towards implicit methods.
Start to change the interface of a method generator (Method class) to
support two different functions: `generate` and `implicit_expression`:

 - `implicit_expression` returns an expression template that specifies
   to the user the format of the implicit expression to expect

 - `generate` replaces `__call__`, except that it takes as a parameter
   a list of *solver hooks*, which are callbacks provided by the user
   that, given a particular instance of an implicit expression, will
   return a new expression for the solve component

In order to facilitate the above changes, passing component id names
is moved to the constructor.

Provide a default solver hook (ScipySolverGenerator) for the Python
backend to demonstrate the use of the implicit tools that have been
developed.

Add tools to simplify the manipulation of expressions, including
parsing and matching.

Bug: #8