Structural extraction: Fix an invalid pattern for If-Then.
As an attempt to be more intelligent, the code tried to swap the role of
"then" and "merge" when it detected that "then" and "merge" were
associated with the "false" and "true" branches of the conditional
instead of the other way around.

This strategy was incomplete, because in order for this to work, the
condition also has to be negated. As a result some incorrect control
flow was being generated (yikes!).

For now, don't implement the advanced strategy. Also fix a test that
implicitly relied on it.

Bug: #51